requirements.txt
# Flask Application Dependencies
# Minecraft Bot Hub Web Server

# Core Flask
Flask>=2.3.0
Flask-SocketIO>=5.3.0
Flask-Session>=0.5.0

# WebSocket and Real-time
python-socketio>=5.8.0
python-engineio>=4.6.0

# HTTP and API
requests>=2.31.0
aiohttp>=3.8.0

# Data handling
PyYAML>=6.0
configparser>=5.3.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-flask>=1.2.0

# Production server
gunicorn>=21.2.0
eventlet>=0.33.0

# Security
cryptography>=41.0.0
bcrypt>=4.0.0

# Database (optional)
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.0.0

# Caching (optional)
Flask-Caching>=2.0.0
redis>=4.5.0

# Background tasks (optional)
celery>=5.3.0
redis>=4.5.0